requirements.txt
gunicorn==23.0.0
Flask==3.1.0
Flask-CORS==5.0.0
# tensorflow==2.10.0
torch==2.5.1
Pillow==11.0.0
numpy==1.24.4